+++ src/Javis.jl
@@ -5,6 +5,7 @@ using LaTeXStrings
 using LightXML
 import Luxor
 import Luxor: Point, @layer
+using ProgressMeter
 using Random
 
 const FRAMES_SYMBOL = [:same]
@@ -1049,7 +1050,7 @@ function javis(
     end
 
     filecounter = 1
-    for frame in frames
+    @showprogress 1 "Rendering frames..." for frame in frames
         background_settings = ActionSetting()
         Drawing(
             video.width,