+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Javis.jl - Changelog
 
 ## Unreleased
+- Ability to scale an object with `Scaling`. Works similar to `Translation` and `Rotation` 
 - Added JuliaFormatter GitHub Action
 
 ## 0.1.2 (24th of August 2020)