+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Show progress of rendering using [ProgressMeter.jl](https://github.com/timholy/ProgressMeter.jl)
 
 ## Unreleased
+- Ability to scale an object with `Scaling`. Works similar to `Translation` and `Rotation` 
 - Added JuliaFormatter GitHub Action
 
 ## 0.1.2 (24th of August 2020)