+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Javis.jl - Changelog
 
+## Unreleased v0.2
+- Show progress of rendering using [ProgressMeter.jl](https://github.com/timholy/ProgressMeter.jl)
+
 ## Unreleased
 - Ability to scale an object with `Scaling`. Works similar to `Translation` and `Rotation` 
 - Added JuliaFormatter GitHub Action